requirements.txt
streamlit== 1.50.0
google-generativeai==0.3.1
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3
Pillow==11.0.0
python-dotenv==1.0.0